+++ core/src/main/java/com/smartsolutions/datwall/repositories/models/IApp.kt
@@ -9,5 +9,5 @@ interface IApp: Parcelable {
     var allowAnnotations: String?
     var blockedAnnotations: String?
 
-    fun accessHashCode(): Int
+    fun accessHashCode(): Long
 }
